refactor and added missing tests
